pyproject.toml
[project]
name = "opendata-tool"
dynamic = ["version"]
description = "Scientific metadata preparation and packaging tool"
authors = [
    {name = "Antigravity", email = "antigravity@example.com"}
]
dependencies = [
    "nicegui>=1.4.0",
    "pydantic[email]>=2.0.0",
    "pyyaml>=6.0.0",
    "google-genai>=1.0.0",
    "google-generativeai>=0.8.0",
    "python-docx>=1.1.0",
    "pypandoc>=1.13",
    "pydicom>=2.4.0",
    "bibtexparser>=1.4.0",
    "h5py>=3.10.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "requests>=2.31.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
opendata = "opendata.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {file = "src/opendata/VERSION"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"opendata" = ["prompts/*.md", "ui/**/*", "VERSION"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "ARG", "ERA", "PL"]
ignore = []
